Add initial version of mungekey(8)
Add the initial version of mungekey, the MUNGE key management utility.
The key is created directly from "/dev/urandom"; this will be revised
in later commits.

The sharness tests are updated to use mungekey.  The rpm spec
file removes key generation, and instead outputs a message to run
mungekey to create a key.  The QUICKSTART documentation has been
updated accordingly.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>