Add mungekey sharnanigans using DEBUG prereq
Add a sharness DEBUG prereq based on the NDEBUG define in "config.h".

Using this DEBUG prereq, add a sharness test to check the
mungekey command-line processing of an unimplemented option.
The GETOPT_DEBUG_SHORT_OPTS define is added to "conf.c".  This
string is appended to the "short_opts" optstring when configured
with --enable-debug.  The new sharness test checks the short-option
specified in GETOPT_DEBUG_SHORT_OPTS for code coverage of the default
case of the switch statement in parse_cmdline().

Also, specify the !DEBUG prereq for the sharness test checking that
nothing is written to stdout or stderr when successfully creating a
key without the --verbose command-line option.  Debug messages could
cause this test to fail.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>