Add log_err_or_warn and refactor file/dir checks
The --force command-line option converts certain errors into warnings
in order to force the daemon to run if at all possible.  However,
the specific errors which can be forced are at times unclear.

Add log_err_or_warn() for logging fatal errors that can be overridden
via the --force command-line option.  If that option is specified,
the fatal error is converted into a non-fatal warning.  The error
message should indicate whether the error can be overridden; that
change will be addressed in a future commit.

Refactor (and fix bugs in) the file and directory checks capable of
being forced accordingly.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>