Add RFC 5869 test cases for HKDF
Add the HKDF test vectors from RFC 5869: HMAC-based Extract-and-Expand
Key Derivation Function (HKDF), Appendix A.

- Test Case 1: SHA-256 basic test
- Test Case 2: SHA-256 with longer inputs/outputs
- Test Case 3: SHA-256 with zero-length salt/info
- Test Case 4: SHA-1 basic test
- Test Case 5: SHA-1 with longer inputs/outputs
- Test Case 6: SHA-1 with zero-length salt/info
- Test Case 7: SHA-1 with no salt and zero-length info

Test case #3 and #6 fail due to an assert() error: "mac_init:
Assertion `keylen > 0' failed".  Consequently, these two cases are
being skipped and will be fixed in a later commit.  Note that with
libtap, todo() cannot be used here since it is for a test expected
to fail with an appropriate error code.  Instead, skip() must be
used since this failure is caused by an assert() which terminates
execution via abort().

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>