Change munged_start_daemon arg-1 processing
The sharness function munged_start_daemon() uses its first argument
(if specified) to exec the daemon.  This is used to run the daemon
under valgrind.  However, this is awkward if additional arguments
need to be passed to the daemon when not running under valgrind,
as it requires the first argument to be "".

Change the argument processing so that iff the first argument starts
with "--exec=", its value will be used to exec the daemon.

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 9.7, 8.11, 7.11, 6.0.10, 5.0.10, 4.0, 3.1
- Fedora Rawhide, 29
- FreeBSD 11.2
- NetBSD 7.1.2
- OpenBSD 6.3
- Ubuntu 18.10, 18.04.1 LTS, 16.04.5 LTS, 14.04.5 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>