Fix sharness check for non-interactive sudo
The sharness tests check if non-interactive sudo is allowed when
setting the SUDO prerequisite.  This check uses the "--non-interactive"
sudo option to prevent prompting for a password.  However, the
long-form of this option is not available in sudo-1.8.5p2 (Debian 7.11)
or sudo-1.8.3p1 (Ubuntu 12.04 LTS).

Fix the check for non-interactive sudo by replacing the
"--non-interactive" long option with the "-n" short option.  The short
option was added in sudo-1.7.0.  As such, it is available for testing
in Debian 6.0.10 (sudo-1.7.4p4).

Reference:
- https://www.sudo.ws/legacy.html#1.7.0

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 9.7, 8.11, 7.11, 6.0.10, 5.0.10, 4.0, 3.1
- Fedora Rawhide, 29
- NetBSD 7.1.2
- OpenBSD 6.3
- Ubuntu 18.10, 18.04.1 LTS, 16.04.5 LTS, 14.04.5 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>