Append trailing comment line to makefile variables
Append a line with an indented comment to Makefile variables.
For variables spanning multiple lines, this allows all lines to
end with a trailing backslash without having to treat the last line
differently in order to avoid the error "blank line following trailing
backslash".  This trailing comment makes it less error-prone to sort
or re-arrange lines within a variable definition.

Tested:
- CentOS 7.6.1810, 6.10
- Debian sid, 9.8, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora Rawhide, 29
- FreeBSD 11.2
- NetBSD 7.1.2
- OpenBSD 6.3
- Ubuntu 18.10, 18.04.2 LTS, 16.04.5 LTS, 14.04.5 LTS, 12.04.5 LTS

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>