Add systemd RuntimeDirectory & RuntimeDirectoryMode
Create the "%{_runstatedir}/munge" directory using the systemd unit
configuration options RuntimeDirectory and RuntimeDirectoryMode.
These options were added in systemd 211.  Back in commit d034b3a,
tmpfiles.d was used instead since CentOS 7 at the time was running
systemd 208; but, CentOS 7.6.1810 is running systemd 219.

The creation of the "tmpfiles.d/munge.conf" file is being kept for
now in case it is still needed somewhere, but this file is no longer
being installed.

The RPM %ghost tag in munge.spec is once again added back to the
"%{_runstatedir}/munge" directory since this directory is now created
by systemd when the munge service is started, and removed by systemd
when the munge service is stopped.

Reference:
- https://www.freedesktop.org/software/systemd/man/systemd.exec.html

Tested:
- CentOS 7.6.1810 (systemd 219)
- Fedora 29 (systemd 239)

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>